(BROKEN) Individual Pages for Projects
Attempt to get individual pages for projects.
Uncommenting the code will allow clicking on projects in dashboard and redirecting to a unique page for the project, but no data loads.

The alert "No routes matched location "/projects/{project_id}" " is displayed in console

The "delete" and "update status" buttons are overlapped
This causes update status to be inaccessable, and a redirect to the recently-removed project's page whenever trying to delete